funcionalidad principal a un 90

+++ app/App/Controllers/Admin/FamiliasController.php
@@ -1,4 +1,4 @@
-<?php 
+<?php
 
 namespace App\Controllers\Admin;
 
@@ -57,7 +57,8 @@ class FamiliasController extends Controller
 		$model->setTable("ma_familias_1");
 		$model->setId("idfamilia");
 
-		$arrData = $model->orderBy("idfamilia", "DESC")->get();
+		$arrData = $model->join("ma_ordenes_1", "idorden")->orderBy("idfamilia", "DESC")->get();
+		// dep($arrData,1);
 		$data = [];
 		$num = 0;
 
@@ -74,7 +75,7 @@ class FamiliasController extends Controller
 			if ($this->permisos['perm_u'] == 1) {
 				$btnEdit = '<a class="dropdown-item" href="javascript:fntEdit(' . $arrData[$i]['idfamilia'] . ');"><i class="bx bx-edit-alt me-1"></i> Editar</a>';
 			}
-
+			$arrData[$i]['num'] = $num;
 			$arrData[$i]['options'] = '<div class="d-flex flex-row"><div class="ms-3 dropdown"><button type="button" class="btn p-0 dropdown-toggle hide-arrow" data-bs-toggle="dropdown"><i class="bx bx-dots-vertical-rounded"></i></button><div class="dropdown-menu">' . $btnEdit . $btnDelete . '</div></div></div>';
 		}
 		return $this->respondWithJson($response, $arrData);
@@ -105,16 +106,14 @@ class FamiliasController extends Controller
 		$model->setTable("ma_familias_1");
 		$model->setId("idfamilia");
 
-		/*
-		$existe = $model->where("field", "LIKE", $data['field'])->first();
+		$existe = $model->where("fam_nombre", "LIKE", $data['fam_nombre'])->where("idorden", $data["idorden"])->first();
 		if (!empty($existe)) {
 			$msg = "Ya tiene un usuario registrado con ese nombre";
 			return $this->respondWithError($response, $msg);
 		}
-		*/
 
 		$rq = $model->create([
-			'idsuborden' => $data['idsuborden'],
+			'idorden' => $data['idorden'],
 			'fam_nombre' => $data['fam_nombre'],
 			// 'fam_date' => $data['fam_date'],
 		]);
@@ -135,15 +134,12 @@ class FamiliasController extends Controller
 	 */
 	public function validar($data)
 	{
-		if (empty($data['idsuborden'])) {
+		if (empty($data['idorden'])) {
 			return false;
 		}
 		if (empty($data['fam_nombre'])) {
 			return false;
 		}
-		// if (empty($data['fam_date'])) {
-		// 	return false;
-		// }
 		return true;
 	}
 
@@ -212,16 +208,14 @@ class FamiliasController extends Controller
 		$model->setTable("ma_familias_1");
 		$model->setId("idfamilia");
 
-		/*
-		$existe = $model->where("field", "LIKE", $data['field'])->first();
+		$existe = $model->where("fam_nombre", "LIKE", $data['fam_nombre'])->where("idfamilia", "!=", $data["idfamilia"])->where("idorden", $data["idorden"])->first();
 		if (!empty($existe)) {
 			$msg = "Ya tiene un usuario registrado con ese nombre";
 			return $this->respondWithError($response, $msg);
 		}
-		*/
 
-		$rq = $model->update($data['idfamilia'],[
-			'idsuborden' => $data['idsuborden'],
+		$rq = $model->update($data['idfamilia'], [
+			'idorden' => $data['idorden'],
 			'fam_nombre' => $data['fam_nombre'],
 			// 'fam_date' => $data['fam_date'],
 		]);
@@ -245,7 +239,7 @@ class FamiliasController extends Controller
 		if (empty($data['idfamilia'])) {
 			return false;
 		}
-		if (empty($data['idsuborden'])) {
+		if (empty($data['idorden'])) {
 			return false;
 		}
 		if (empty($data['fam_nombre'])) {
@@ -276,6 +270,15 @@ class FamiliasController extends Controller
 
 		$rq = $model->find($data['idfamilia']);
 		if (!empty($rq)) {
+			// consultar a la tabla ma_subfamilias_1 si existe algun registro con el idfamilia
+			$model2 = new TableModel;
+			$model2->setTable("ma_subfamilias_1");
+			$model2->setId("idsubfamilia");
+			$existe = $model2->where("idfamilia", "=", $data['idfamilia'])->first();
+			if (!empty($existe)) {
+				$msg = "No se puede eliminar el registro, existen subfamilias asociadas";
+				return $this->respondWithError($response, $msg);
+			}
 			$rq = $model->delete($data["idfamilia"]);
 			if (!empty($rq)) {
 				$msg = "Datos eliminados correctamente";
@@ -285,5 +288,16 @@ class FamiliasController extends Controller
 
 		$msg = "Error al eliminar los datos";
 		return $this->respondWithError($response, $msg);
+	}
+
+	/**
+	 * Lista de ordenes activos
+	 */
+	public function ordenes($request, $response)
+	{
+		$model = new TableModel;
+		$model->setTable("ma_ordenes_1");
+		$model->setId("idorden");
+		return $this->respondWithJson($response, $model->orderBy("idorden", "DESC")->get());
 	}
 }
